requirements.txt
streamlit>=1.33
pandas>=2.1
numpy>=1.26
matplotlib>=3.8
python-dotenv>=1.0
openpyxl>=3.1
openai>=1.30